backend/requirements.txt
azure-functions
azure-cosmos
PyJWT
bcrypt
python-dotenv
requests